Modify UserManual to be more responsive
Take away the fixed navbar, instead add a fixed
icon to the top left of the User Manual page,
which expands a side navigation bar when clicked.

Introduce a Preamble section to the User Manual,
which clarifies some things about using the User Manual.

Additionally, add some CSS for the User Manual and various
classes to App.css

Signed-off-by: Kevin Morris <kmorris@zerocost.dev>